+++ src/main/java/com/microej/demo/widget/progressbar/widget/ProgressBar.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2023 MicroEJ Corp. All rights reserved.
+ * Copyright 2015-2024 MicroEJ Corp. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
 package com.microej.demo.widget.progressbar.widget;
@@ -58,7 +58,7 @@ public class ProgressBar extends Widget {
 
 		// draw percentage string
 		int percentage = Math.round(PROGRESS_MULTIPLIER * this.progress);
-		String string = Integer.toString(percentage) + "%"; //$NON-NLS-1$
+		String string = percentage + "%"; //$NON-NLS-1$
 		g.setColor(getStringColor(style));
 		StringPainter.drawStringInArea(g, string, style.getFont(), 0, 0, contentWidth, contentHeight,
 				style.getHorizontalAlignment(), style.getVerticalAlignment());