+++ src/main/java/com/microej/demo/widget/sliderwithvalue/widget/SliderWithValue.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2023 MicroEJ Corp. All rights reserved.
+ * Copyright 2021-2024 MicroEJ Corp. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
 package com.microej.demo.widget.sliderwithvalue.widget;
@@ -181,6 +181,8 @@ public class SliderWithValue extends Widget implements EventHandler {
 			case Buttons.PRESSED:
 				onPointerPressed();
 				return true;
+			default:
+				break;
 			}
 		}
 		return super.handleEvent(event);